refactor: keep environment and SSROptions inside render
