+++ packages/astro/src/core/render/environment.ts
@@ -1,12 +1,13 @@
 import type { MarkdownRenderingOptions } from '@astrojs/markdown-remark';
-import type { RuntimeMode, SSRLoadedRenderer } from '../../@types/astro';
+import type { RuntimeMode, SSRLoadedRenderer, AstroSettings } from '../../@types/astro';
 import type { LogOptions } from '../logger/core.js';
 import type { RouteCache } from './route-cache.js';
+import type { ModuleLoader } from '../module-loader';
 
 /**
  * An environment represents the static parts of rendering that do not change
  * between requests. These are mostly known when the server first starts up and do not change.
- * Thus they can be created once and passed through to renderPage on each request.
+ * Thus, they can be created once and passed through to renderPage on each request.
  */
 export interface Environment {
 	/**
@@ -42,3 +43,8 @@ export type CreateEnvironmentArgs = Environment;
 export function createEnvironment(options: CreateEnvironmentArgs): Environment {
 	return options;
 }
+
+export type DevelopmentEnvironment = Environment & {
+	loader: ModuleLoader;
+	settings: AstroSettings;
+};
