+++ packages/astro/src/prerender/routing.ts
@@ -1,7 +1,8 @@
 import type { AstroSettings, ComponentInstance, RouteData } from '../@types/astro';
 import { RedirectComponentInstance, routeIsRedirect } from '../core/redirects/index.js';
-import { preload, type DevelopmentEnvironment } from '../vite-plugin-astro-server/index.js';
+import { preload } from '../vite-plugin-astro-server/index.js';
 import { getPrerenderStatus } from './metadata.js';
+import type { DevelopmentEnvironment } from '../core/render';
 
 type GetSortedPreloadedMatchesParams = {
 	env: DevelopmentEnvironment;
