+++ packages/astro/src/vite-plugin-astro-server/environment.ts
@@ -2,14 +2,10 @@ import { createResolve } from './resolve.js';
 import type { ModuleLoader } from '../core/module-loader';
 import type { AstroSettings, RuntimeMode, SSRManifest } from '../@types/astro.js';
 import type { LogOptions } from '../core/logger/core.js';
-import { createEnvironment, type Environment } from '../core/render/index.js';
 import { RouteCache } from '../core/render/route-cache.js';
 import { isServerLikeOutput } from '../prerender/utils.js';
-
-export type DevelopmentEnvironment = Environment & {
-	loader: ModuleLoader;
-	settings: AstroSettings;
-};
+import type { DevelopmentEnvironment } from '../core/render';
+import { createEnvironment } from '../core/render/index.js';
 
 export function createDevelopmentEnvironment(
 	manifest: SSRManifest,
