+++ packages/astro/src/vite-plugin-astro-server/route.ts
@@ -13,11 +13,16 @@ import { AstroErrorData, isAstroError } from '../core/errors/index.js';
 import { warn } from '../core/logger/core.js';
 import { loadMiddleware } from '../core/middleware/loadMiddleware.js';
 import { getStylesForURL } from './css.js';
-import type { DevelopmentEnvironment } from './environment';
-import { preload, type SSROptions } from './index.js';
+import { preload } from './index.js';
 import { getComponentMetadata } from './metadata.js';
 import { getScriptsForURL } from './scripts.js';
-import { createRenderContext, getParamsAndProps, tryRenderRoute } from '../core/render/index.js';
+import {
+	createRenderContext,
+	type DevelopmentEnvironment,
+	type SSROptions,
+	getParamsAndProps,
+	tryRenderRoute,
+} from '../core/render/index.js';
 import { createRequest } from '../core/request.js';
 import { matchAllRoutes } from '../core/routing/index.js';
 import { isPage, resolveIdToUrl, viteID } from '../core/util.js';
